app/requirements.txt
python-dotenv
Flask # Opcional, para un futuro webhook. Por ahora, simularemos la entrada.